(flask) Finish prototype

+++ flask/application/jobs/cleanup_job.py
@@ -1,16 +1,25 @@
 import datetime
 import os
-from . import BaseJob
 from application.models import JobModel
+from .jobs import BaseJob
 
 
 class CleanupJob(BaseJob):
     SCHEDULE = os.environ.get('CLEANUP_JOB_CRON', '0 0 * * 0')
     DAYS_AGO = int(os.environ.get('CLEANUP_JOB_DAYS', '3'))
 
-    def _run(self, *args):
+    def _run(self, *_):
+        self.job_model.result_link = '/jobs'  # url_for('jobs.index')
+
         days_ago = datetime.datetime.now() - datetime.timedelta(days=CleanupJob.DAYS_AGO)
-        old_jobs = JobModel.query.filter(updated_at__lt=days_ago)
-        # old_jobs._raw_delete(old_jobs.db)
+        self.logger.info("Removing old job before '%s'...", days_ago)
+        old_jobs = self.session.query(JobModel).filter(
+            JobModel.updated_at <= days_ago)
+        count = old_jobs.count()
+        if count < 1:
+            self.logger.info('No obsolete jobs')
+            return
+
         old_jobs.delete()
-        # TODO: commit?
+        self.session.commit()
+        self.logger.info("Deleted '%s' old jobs", count)
